refactor(server)!: split model providers into two REST collections
Review follow-up to the endpoint added in the previous commit.

`GET /v1/model_providers` returned built-in provider families and custom
providers in one discriminated list, while #12264/#12265/#12266 plan
POST /model_providers and PATCH|DELETE /model_providers/{id} against custom
providers only. That leaves a collection whose members are not uniformly
addressable: the 14 built-in families have no id, no URI, and reject every
write, and `limit` under-reports page size because built-ins are prepended
to the first page outside the cursor window.

Splits them into two homogeneous collections, each named for what it holds:

- `GET /v1/model_providers` -- built-in families. A fixed enum narrowed by
  PHOENIX_ALLOWED_PROVIDERS, so it returns a plain `ResponseBody` with no
  cursor, and touches no database.
- `GET /v1/custom_model_providers` -- rows in
  `generative_model_custom_providers`, paginated, where cursor/limit/
  next_cursor now mean what they mean on every other v1 collection.

Both keep their `{id}` slot free, so the mutating siblings can land on
/v1/custom_model_providers/{id} without a static segment competing with a
provider id. Flat compound collections match the surrounding surface
(`dataset_labels`, `prompt_versions`, `span_annotations`), and both modules
share the `model_providers` OpenAPI tag so they still group in the docs, the
way `dataset_labels` sits under `datasets`.

Because each collection is homogeneous, no discriminated union is needed.
The built-in entry's provider field is `provider: ModelProvider`, matching
`BuiltInProviderModelSelection` in `agents/model_selection.py`, so a listed
provider can be passed straight to /v1/agents/chat without renaming fields.

Also folds in the rest of the review:

- Validates the cursor's GlobalID node type via the existing
  `from_global_id_with_expected_type`. A cursor minted by another endpoint
  (e.g. `Project:5` from GET /v1/projects) was previously used verbatim as a
  `generative_model_custom_providers.id`, returning a wrong page with 200
  instead of 422.
- Hoists cursor parsing and response serialization out of the DB session
  block; only the query needs a session.
- Deduplicates the PHOENIX_ALLOWED_PROVIDERS visibility policy into
  `PLAYGROUND_CLIENT_REGISTRY.list_allowed_providers()`, now called by both
  this router and `Query.model_providers`. The two surfaces previously read
  the allow-list through different plumbing and could drift.
- Registers both routes in the integration suite's endpoint-coverage
  tuples. `tests/integration/_helpers.py` asserts at import time that every
  mounted route is listed, so the missing entry took down conftest and
  failed the whole integration suite with exit 4 before any test ran.

BREAKING CHANGE: GET /v1/model_providers no longer returns custom providers
or a `next_cursor`, and built-in entries expose `provider` rather than
`kind` + `provider_key`. Custom providers move to GET
/v1/custom_model_providers. The endpoint is unreleased, so no published
client is affected.
