+++ schemas/openapi.json
@@ -7014,9 +7014,41 @@
         "tags": [
           "model_providers"
         ],
-        "summary": "List all model providers",
-        "description": "Retrieve the built-in model provider families along with a paginated list of user-defined custom providers. Built-in families are a fixed set and are only returned on the first page, i.e. when no `cursor` is supplied; `cursor`, `next_cursor`, and `limit` apply exclusively to the custom-provider portion of the list. Encrypted custom-provider credentials are never returned.",
+        "summary": "List built-in model provider families",
+        "description": "Retrieve the built-in model provider families available to this deployment. Built-in families are a fixed enum rather than stored records, so this list is not paginated; it is narrowed by the PHOENIX_ALLOWED_PROVIDERS environment variable when that is set. User-defined providers are listed separately by `GET /v1/custom_model_providers`.",
         "operationId": "getModelProviders",
+        "responses": {
+          "200": {
+            "description": "A list of built-in model provider families",
+            "content": {
+              "application/json": {
+                "schema": {
+                  "$ref": "#/components/schemas/GetModelProvidersResponseBody"
+                }
+              }
+            }
+          },
+          "403": {
+            "description": "Forbidden",
+            "content": {
+              "text/plain": {
+                "schema": {
+                  "type": "string"
+                }
+              }
+            }
+          }
+        }
+      }
+    },
+    "/v1/custom_model_providers": {
+      "get": {
+        "tags": [
+          "model_providers"
+        ],
+        "summary": "List custom model providers",
+        "description": "Retrieve a paginated list of user-defined custom model providers. Encrypted provider credentials are never returned. Built-in provider families are listed separately by `GET /v1/model_providers`.",
+        "operationId": "getCustomModelProviders",
         "parameters": [
           {
             "name": "cursor",
@@ -7052,11 +7084,11 @@
         ],
         "responses": {
           "200": {
-            "description": "A list of model providers with pagination information",
+            "description": "A list of custom model providers with pagination information",
             "content": {
               "application/json": {
                 "schema": {
-                  "$ref": "#/components/schemas/GetModelProvidersResponseBody"
+                  "$ref": "#/components/schemas/GetCustomModelProvidersResponseBody"
                 }
               }
             }
@@ -10442,16 +10474,9 @@
       },
       "BuiltInModelProvider": {
         "properties": {
-          "kind": {
-            "type": "string",
-            "const": "builtin",
-            "title": "Kind",
-            "default": "builtin"
-          },
-          "provider_key": {
-            "type": "string",
-            "title": "Provider Key",
-            "description": "The stable key identifying the built-in provider family (e.g. 'OPENAI')."
+          "provider": {
+            "$ref": "#/components/schemas/ModelProvider",
+            "description": "The provider family identifier, accepted wherever a built-in model provider is specified (e.g. 'OPENAI')."
           },
           "name": {
             "type": "string",
@@ -10461,7 +10486,7 @@
         },
         "type": "object",
         "required": [
-          "provider_key",
+          "provider",
           "name"
         ],
         "title": "BuiltInModelProvider"
@@ -12111,12 +12136,6 @@
       },
       "CustomModelProvider": {
         "properties": {
-          "kind": {
-            "type": "string",
-            "const": "custom",
-            "title": "Kind",
-            "default": "custom"
-          },
           "id": {
             "type": "string",
             "title": "Id",
@@ -12142,7 +12161,7 @@
           "provider": {
             "type": "string",
             "title": "Provider",
-            "description": "The provider label of the custom provider."
+            "description": "The free-form provider label recorded on the custom provider. Unlike the `provider` of a built-in family, this is not drawn from a fixed set."
           },
           "sdk": {
             "type": "string",
@@ -13658,6 +13677,34 @@
         ],
         "title": "GetApiKeysResponseBody"
       },
+      "GetCustomModelProvidersResponseBody": {
+        "properties": {
+          "data": {
+            "items": {
+              "$ref": "#/components/schemas/CustomModelProvider"
+            },
+            "type": "array",
+            "title": "Data"
+          },
+          "next_cursor": {
+            "anyOf": [
+              {
+                "type": "string"
+              },
+              {
+                "type": "null"
+              }
+            ],
+            "title": "Next Cursor"
+          }
+        },
+        "type": "object",
+        "required": [
+          "data",
+          "next_cursor"
+        ],
+        "title": "GetCustomModelProvidersResponseBody"
+      },
       "GetDatasetLabelResponseBody": {
         "properties": {
           "data": {
@@ -13782,41 +13829,15 @@
         "properties": {
           "data": {
             "items": {
-              "oneOf": [
-                {
-                  "$ref": "#/components/schemas/BuiltInModelProvider"
-                },
-                {
-                  "$ref": "#/components/schemas/CustomModelProvider"
-                }
-              ],
-              "discriminator": {
-                "propertyName": "kind",
-                "mapping": {
-                  "builtin": "#/components/schemas/BuiltInModelProvider",
-                  "custom": "#/components/schemas/CustomModelProvider"
-                }
-              }
+              "$ref": "#/components/schemas/BuiltInModelProvider"
             },
             "type": "array",
             "title": "Data"
-          },
-          "next_cursor": {
-            "anyOf": [
-              {
-                "type": "string"
-              },
-              {
-                "type": "null"
-              }
-            ],
-            "title": "Next Cursor"
           }
         },
         "type": "object",
         "required": [
-          "data",
-          "next_cursor"
+          "data"
         ],
         "title": "GetModelProvidersResponseBody"
       },
