+++ src/phoenix/server/api/routers/v1/__init__.py
@@ -11,6 +11,7 @@ from .annotation_configs import router as annotation_configs_router
 from .annotations import router as annotations_router
 from .api_keys import router as api_keys_router
 from .chat_completions import router as chat_completions_router
+from .custom_model_providers import router as custom_model_providers_router
 from .dataset_labels import router as dataset_labels_router
 from .datasets import router as datasets_router
 from .documents import router as documents_router
@@ -74,6 +75,7 @@ def create_v1_router(authentication_enabled: bool) -> APIRouter:
     viewer_restricted_router.include_router(prompts_router)
     viewer_restricted_router.include_router(projects_router)
     viewer_restricted_router.include_router(model_providers_router)
+    viewer_restricted_router.include_router(custom_model_providers_router)
     viewer_restricted_router.include_router(sessions_router)
     viewer_restricted_router.include_router(documents_router)
     viewer_restricted_router.include_router(users_router)
